move function to test file
