+++ lib/log-processor/utils/index.test.mjs
@@ -2,7 +2,6 @@ import assert from 'node:assert';
 import fs from 'node:fs';
 
 import RouteMetricsResults from './index.mjs';
-import read from './read-log.mjs';
 
 const files = [
   {name: 'minimal-route-metrics.log', counts: {
@@ -35,20 +34,20 @@ describe('RouteMetricsResults', function() {
     const {lines, header, patch, status, proc, eventloop, gc, route} = counts;
     const file = `./test/log-files/${name}`;
     it(`${name}: read an array of log lines`, async function() {
-      const logLines = await read(file);
+      const logLines = await readRouteMetricsLog(file);
       // captured with wc -l
       assert.equal(logLines.length, lines);
       logLines.forEach(line => assert.equal(typeof line, 'string'));
     });
 
     it(`${name}: should convert lines to logObjects`, async function() {
-      const logObjects = await read(file, {convert: true});
+      const logObjects = await readRouteMetricsLog(file, {convert: true});
       assert.equal(logObjects.length, lines);
       logObjects.forEach(line => assert.equal(typeof line, 'object'));
     });
 
     it(`${name}: indexes an array of logObjects`, async function() {
-      const logObjects = await read(file, {convert: true});
+      const logObjects = await readRouteMetricsLog(file, {convert: true});
       const rmr = new RouteMetricsResults();
       rmr.indexArray(logObjects);
 
@@ -225,3 +224,30 @@ POST http://127.0.0.1:4000/api/users/login,success,1
 
 });
 
+//
+// the following will need to move to a test/utils (or similar) place if more
+// tests end up using it.
+//
+
+import fsp from 'node:fs/promises';
+
+//
+// take the brute force approach now, and read the entire file into memory. if
+// it ever becomes a problem, we should just request bigger machines as it's
+// much more complex to handle all this with streaming data.
+//
+async function readRouteMetricsLog(filename, options = {}) {
+  const {convert = false} = options;
+
+  // read the route-metrics log file
+  let logText = await fsp.readFile(filename, {encoding: 'utf8'});
+  logText = logText.split('\n').slice(0, -1);
+
+  if (!convert) {
+    return logText;
+  }
+
+  const logObjects = logText.map(JSON.parse);
+
+  return logObjects;
+}
